+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - Request methods
+- PixReversal description
 
 ## [0.11.3] - 2024-06-19
 ### Fixed